Moved stakeNewLPTokens to helpers
